+++ Cargo.toml
@@ -2,7 +2,6 @@
 
 members = [
   "tokio",
-  "tokio-codec",
   "tokio-executor",
   "tokio-fs",
   "tokio-io",
@@ -12,5 +11,8 @@ members = [
   "tokio-test",
   "tokio-timer",
   "tokio-tls",
+  "tokio-util",
+
+  # Crates used by tests
   "build-tests",
 ]
