+++ ci/patch.toml
@@ -2,7 +2,6 @@
 # repository.
 [patch.crates-io]
 tokio = { path = "tokio" }
-tokio-codec = { path = "tokio-codec" }
 tokio-executor = { path = "tokio-executor" }
 tokio-fs = { path = "tokio-fs" }
 tokio-io = { path = "tokio-io" }
@@ -11,3 +10,4 @@ tokio-net = { path = "tokio-net" }
 tokio-sync = { path = "tokio-sync" }
 tokio-timer = { path = "tokio-timer" }
 tokio-tls = { path = "tokio-tls" }
+tokio-util = { path = "tokio-util" }
