+++ tokio-util/src/codec/bytes_codec.rs
@@ -1,5 +1,6 @@
-use crate::decoder::Decoder;
-use crate::encoder::Encoder;
+use crate::codec::decoder::Decoder;
+use crate::codec::encoder::Encoder;
+
 use bytes::{BufMut, Bytes, BytesMut};
 use std::io;
