+++ tokio-util/src/lib.rs
@@ -1,0 +1,17 @@
+#![doc(html_root_url = "https://docs.rs/tokio-util/0.2.0-alpha.6")]
+#![warn(
+    missing_debug_implementations,
+    missing_docs,
+    rust_2018_idioms,
+    unreachable_pub
+)]
+#![deny(intra_doc_link_resolution_failure)]
+#![doc(test(
+    no_crate_inject,
+    attr(deny(warnings, rust_2018_idioms), allow(dead_code, unused_variables))
+))]
+
+//! Utilities for working with Tokio.
+
+pub mod codec;
+pub mod udp;
