+++ tokio-util/src/udp/frame.rs
@@ -1,14 +1,14 @@
-use super::UdpSocket;
+use crate::codec::{Decoder, Encoder};
 
-use tokio_codec::{Decoder, Encoder};
+use tokio::net::UdpSocket;
 
 use bytes::{BufMut, BytesMut};
-use core::task::{Context, Poll};
 use futures_core::{ready, Stream};
 use futures_sink::Sink;
 use std::io;
 use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
 use std::pin::Pin;
+use std::task::{Context, Poll};
 
 /// A unified `Stream` and `Sink` interface to an underlying `UdpSocket`, using
 /// the `Encoder` and `Decoder` traits to encode and decode frames.
@@ -47,22 +47,17 @@ impl<C: Decoder + Unpin> Stream for UdpFramed<C> {
 
         let (_n, addr) = unsafe {
             // Read into the buffer without having to initialize the memory.
-            let res = ready!(Pin::new(&mut pin.socket).poll_recv_from_priv(cx, pin.rd.bytes_mut()));
+            let res = ready!(Pin::new(&mut pin.socket).poll_recv_from(cx, pin.rd.bytes_mut()));
             let (n, addr) = res?;
             pin.rd.advance_mut(n);
             (n, addr)
         };
-
-        let span = trace_span!("decoding", from.addr = %addr, dgram.length = _n);
-        let _e = span.enter();
-        trace!("trying to decode a frame...");
 
         let frame_res = pin.codec.decode(&mut pin.rd);
         pin.rd.clear();
         let frame = frame_res?;
         let result = frame.map(|frame| Ok((frame, addr))); // frame -> (frame, addr)
 
-        trace!("frame decoded from buffer");
         Poll::Ready(result)
     }
 }
@@ -84,16 +79,11 @@ impl<C: Encoder + Unpin> Sink<(C::Item, SocketAddr)> for UdpFramed<C> {
     fn start_send(self: Pin<&mut Self>, item: (C::Item, SocketAddr)) -> Result<(), Self::Error> {
         let (frame, out_addr) = item;
 
-        let span = trace_span!("sending", to.addr = %out_addr);
-        let _e = span.enter();
-        trace!("encoding frame...");
-
         let pin = self.get_mut();
 
         pin.codec.encode(frame, &mut pin.wr)?;
         pin.out_addr = out_addr;
         pin.flushed = false;
-        trace!(message = "frame encoded", frame.length = pin.wr.len());
 
         Ok(())
     }
@@ -110,17 +100,11 @@ impl<C: Encoder + Unpin> Sink<(C::Item, SocketAddr)> for UdpFramed<C> {
             ..
         } = *self;
 
-        let span = trace_span!("flushing", to.addr = %out_addr, frame.length = wr.len());
-        let _e = span.enter();
-        trace!("flushing frame...");
-
-        let n = ready!(socket.poll_send_to_priv(cx, &wr, &out_addr))?;
+        let n = ready!(socket.poll_send_to(cx, &wr, &out_addr))?;
 
         let wrote_all = n == self.wr.len();
         self.wr.clear();
         self.flushed = true;
-
-        trace!(written.length = n, written.complete = wrote_all);
 
         let res = if wrote_all {
             Ok(())
