+++ tokio-util/tests/framed_write.rs
@@ -1,9 +1,9 @@
 #![warn(rust_2018_idioms)]
 
-use tokio_codec::{Encoder, FramedWrite};
 use tokio_io::AsyncWrite;
 use tokio_test::assert_ready;
 use tokio_test::task::MockTask;
+use tokio_util::codec::{Encoder, FramedWrite};
 
 use bytes::{BufMut, BytesMut};
 use futures_sink::Sink;
