+++ tokio-util/tests/length_delimited.rs
@@ -1,12 +1,12 @@
 #![warn(rust_2018_idioms)]
 
-use tokio::codec::*;
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::prelude::*;
 use tokio_test::task::MockTask;
 use tokio_test::{
     assert_err, assert_ok, assert_pending, assert_ready, assert_ready_err, assert_ready_ok,
 };
+use tokio_util::codec::*;
 
 use bytes::{BufMut, Bytes, BytesMut};
 use futures_util::pin_mut;
